# config/openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation - MiApp
  version: 1.1.0
  description: |
    Documentación intermedia de la API. Esta especificación se mantiene manualmente.
    Incluye endpoints de autenticación, usuarios y tareas con modelos de petición y respuesta clave.
servers:
  - url: http://127.0.0.1:8000
    description: Local Dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCreateRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        name: { type: string }
        roles:
          type: array
          items: { type: string }
    UserUpdateRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
        roles:
          type: array
          items: { type: string }
        name: { type: string }
    UserRolesUpdateRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items: { type: string }
    UserStatusPatchRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        roles:
          type: array
          items: { type: string }
        active: { type: boolean }
        deletedAt: { type: string, format: date-time, nullable: true }
...
paths:
...
  /api/profile:
    get:
      tags: [Auth]
      summary: Perfil del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Perfil, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { description: No autenticado }
  /api/users:
    get:
      tags: [Usuarios]
      summary: Listar usuarios
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: email
          schema:
            type: string
          description: Filtro parcial por email
      responses:
        '200': { description: Lista de usuarios, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }
        '401': { description: No autenticado }
    post:
      tags: [Usuarios]
      summary: Crear usuario (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201': { description: Creado, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { description: Datos inválidos }
        '409': { description: Email duplicado }
  /api/users/{id}:
    put:
      tags: [Usuarios]
      summary: Actualizar usuario (admin, reemplazo completo)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200': { description: Actualizado, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '404': { description: No encontrado }
    delete:
      tags: [Usuarios]
      summary: Eliminar (desactivar) usuario (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200': { description: Eliminado lógicamente }
        '404': { description: No encontrado }
  /api/users/{id}/roles:
    post:
      tags: [Usuarios]
      summary: Actualizar roles (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRolesUpdateRequest'
      responses:
        '200': { description: Roles actualizados }
        '400': { description: Formato inválido }
        '404': { description: No encontrado }
  /api/users/{id}/status:
    patch:
      tags: [Usuarios]
      summary: Activar / desactivar usuario (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserStatusPatchRequest'
      responses:
        '200': { description: Estado actualizado }
        '404': { description: No encontrado }
...
security:
  - bearerAuth: []